move AnchorLink to v1/blocks since it's not atomic element
