+++ cfgov/v1/blocks.py
@@ -1,5 +1,8 @@
+from django.core.exceptions import ValidationError
 from wagtail.wagtailcore import blocks
 from django.utils.module_loading import import_string
+from django.utils.text import slugify
+from .util.util import get_unique_id
 
 
 class AbstractFormBlock(blocks.StructBlock):
@@ -30,3 +33,33 @@ class AbstractFormBlock(blocks.StructBlock):
         handler = None
         method = 'POST'
         icon = 'form'
+
+class AnchorLink(blocks.StructBlock):
+    link_id = blocks.CharBlock(required=False)
+    # full_url = blocks.CharBlock(required=False, default=slugurl)
+
+    def clean(self, data):
+        error_dict = {}
+
+        def format_id(string):
+            words = string.split('_');
+            suffix = ''
+            if string:
+                suffix = '_'
+            if 'anchor' in words:
+                return slugify(string)
+            else:
+                return get_unique_id('anchor_' + slugify(string) + suffix)
+
+        if data:
+            try:
+                data['link_id'] = format_id(data['link_id'])
+            except ValidationError as e:
+                error_dict.update(e.params)
+
+        return super(AnchorLink, self).clean(data)
+
+    class Meta:
+        icon = 'link'
+        template = '_includes/atoms/anchor-link.html'
+        label = 'Anchor link'
